moved OneOfRule to OneOfTypeRule
