+++ jsonschema2pojo-core/src/main/java/org/jsonschema2pojo/rules/SchemaRule.java
@@ -70,7 +70,7 @@ public class SchemaRule implements Rule<JClassContainer, JType> {
         if (schemaNode.has("enum")) {
             javaType = ruleFactory.getEnumRule().apply(nodeName, schemaNode, generatableType, schema);
         } else if(schemaNode.has("oneOf")) {
-            javaType = ruleFactory.getOneOfRule().apply(nodeName, schemaNode, generatableType.getPackage(), schema);
+            javaType = ruleFactory.getOneOfTypeRule().apply(nodeName, schemaNode, generatableType.getPackage(), schema);
         } else {
             javaType = ruleFactory.getTypeRule().apply(nodeName, schemaNode, generatableType.getPackage(), schema);
         }
